third_party/blink/web_tests/storage/indexeddb/mozilla/add-twice-failure-expected.txt

Test IndexedDB behavior adding the same property twice

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

dbname = "add-twice-failure.html"
indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)
objectStore = db.createObjectStore('foo');
key = 10
request = objectStore.add({}, key);
PASS request.result is key
request = objectStore.add({}, key);
PASS event.target.result is undefined.
PASS event.target.error is non-null.
PASS event.target.error.name is 'ConstraintError'
PASS successfullyParsed is true

TEST COMPLETE
